refactor: renamed the routers and procedures to match the new structure
