+++ package.json
@@ -20,7 +20,6 @@
     "@clerk/nextjs": "^4.29.3",
     "@ducanh2912/next-pwa": "^10.2.5",
     "@hookform/resolvers": "^3.3.3",
-    "@next/font": "^14.1.0",
     "@radix-ui/react-alert-dialog": "^1.0.5",
     "@radix-ui/react-avatar": "^1.0.4",
     "@radix-ui/react-checkbox": "^1.0.4",
