+++ src/server/api/routers/company/company.router.ts
@@ -1,14 +1,12 @@
-import { createCompanyRouter } from "@/server/api/routers/company/company-create"
-import { getCompanyRouter } from "@/server/api/routers/company/company-get"
-import { listCompanyRouter } from "@/server/api/routers/company/company-list"
-import { updateCompanyRouter } from "@/server/api/routers/company/company-update"
-import { archiveConferenceRouter } from "@/server/api/routers/conference/conference-archive"
+import { createCompanyProcedure } from "@/server/api/routers/company/company-create.procedure"
+import { getCompanyProcedure } from "@/server/api/routers/company/company-get.procedure"
+import { listCompanyProcedure } from "@/server/api/routers/company/company-list.procedure"
+import { updateCompanyProcedure } from "@/server/api/routers/company/company-update.procedure"
 import { createTRPCRouter } from "@/server/api/trpc"
 
 export const companyRouter = createTRPCRouter({
-  list: listCompanyRouter,
-  get: getCompanyRouter,
-  create: createCompanyRouter,
-  update: updateCompanyRouter,
-  archive: archiveConferenceRouter,
+  list: listCompanyProcedure,
+  get: getCompanyProcedure,
+  create: createCompanyProcedure,
+  update: updateCompanyProcedure,
 })
