+++ src/server/api/routers/user/user-role.procedure.ts
@@ -1,0 +1,8 @@
+import { UserRole } from "@/contracts/user/user-role"
+import { protectedProcedure } from "@/server/api/trpc"
+
+export const getUserRoleProcedure = protectedProcedure.query(
+  async ({ ctx }): Promise<UserRole> => {
+    return (ctx.user.role as UserRole) ?? UserRole.user
+  },
+)
